Refactors pattern loading logic
Begin by assuming that everything is a path, but we want to load it as a
module. So, take the path and after ensuring it is not a ../ kind of path,
convert it to a module name (dots as the separator), then import it with the
builtin `__import__`

This change means we don't need the `module_like` check any longer, so we don't
need to import the `re` module, and since we are using `__import__` we also
don't need to import the `imp` or `importlib` modules!